Generate and/or plot a sample from the prior distribution of M-spline hazard curves

Generates and/or plots the hazard curves (as functions of time) implied by a prior mean for the spline coefficients (a constant hazard by default) and particular priors for the baseline log hazard and smoothness variance

Usage

mspline_priorpred(
  iknots = NULL,
  bknots = c(0, 10),
  df = 10,
  degree = 3,
  coefs_mean = NULL,
  prior_smooth = p_gamma(2, 1),
  prior_loghaz = NULL,
  prior_loghr = NULL,
  x = NULL,
  X = NULL,
  X0 = NULL,
  nonprop = FALSE,
  prior_sdnp = p_gamma(2, 1),
  tmin = 0,
  tmax = 10,
  nsim = 10
)

plot_mspline_priorpred(
  iknots = NULL,
  bknots = c(0, 10),
  df = 10,
  degree = 3,
  coefs_mean = NULL,
  prior_smooth = p_gamma(2, 1),
  prior_loghaz = p_normal(0, 20),
  prior_loghr = NULL,
  x = NULL,
  X = NULL,
  nonprop = FALSE,
  prior_sdnp = p_gamma(2, 1),
  tmin = 0,
  tmax = 10,
  nsim = 10
)

Arguments

iknots

Vector of internal knot locations. If not supplied, df has to be specified, in which case the default is df - degree - 1 equally spaced knots between the boundary knots.

bknots

Vector with two elements, giving the boundary knot locations

df

Desired number of basis terms, or "degrees of freedom" in the spline. If iknots is not supplied, the number of internal knots is then chosen to satisfy this.

degree

Spline polynomial degree (defaults to 3)

coefs_mean

Spline basis coefficients that define the prior mean for the hazard function. By default, these are set to values that define a constant hazard function. They are normalised to sum to 1 internally (if they do not already).

prior_smooth

Gamma prior for the smoothing standard deviation, in models where this is estimated. This should be a call to p_gamma(). The default is p_gamma(2,1).

prior_loghaz

Prior for the baseline log rate parameter (log(eta)), after centering around the log crude event rate in the data. This should be a call to a prior constructor function, such as p_normal(0,1) or p_t(0,2,2). Supported prior distribution families are normal (parameters mean and SD) and t distributions (parameters location, scale and degrees of freedom). The default is a normal distribution with mean 0 and standard deviation 20.

"Baseline" is defined by the continuous covariates taking a value of zero and factor covariates taking their reference level. To use a different baseline, the data should be transformed appropriately beforehand, so that a value of zero has a different meaning.

For continuous covariates, it helps for both computation and interpretation to define the value of zero to denote a typical value in the data, e.g. the mean.

prior_loghr

Priors for log hazard ratios. This should be a call to p_normal() or p_t(). A list of calls can also be provided, to give different priors to different coefficients, where the name of each list component matches the name of the coefficient, e.g. list("age45-59" = p_normal(0,1), "age60+" = p_t(0,2,3))

The default is p_normal(0,2.5) for all coefficients.

x

Only required for models with covariates. A list with components ncovs (number of covariates) and xnames (names of these covariates).

X

Either a list, vector or one-row matrix with one covariate value. Factors must be supplied as multiple numeric (0/1) contrasts. (TODO doc on how to build the names - vignette)

X0

An alternative covariate value (optional). The hazard ratio between the hazards at X and X0 will also be returned.

nonprop

Is the model a non-proportional hazards model (TRUE or FALSE).

prior_sdnp

Prior for the standard deviation parameters that smooth the non-proportionality effects over time in non-proportional hazards models. This should be a call to p_gamma() or a list of calls to p_gamma() with one component per covariate, as in prior_loghr.

tmin

Minimum plotting time. Defaults to lower boundary knot.

tmax

Maximum plotting time. Defaults to upper boundary knot.

nsim

Number of simulations to draw

Value

mspline_priorpred returns a data frame of the samples, and plot_mspline_priorpred generates a plot.