Draw a sample from the prior distribution of the parameters governing the hazard in a survextrap model for given covariates.

This is used, for example, in mspline_priorpred to visualise the prior distribution around hazard curves implied by a particular M-spline model and parameter priors.

Usage

prior_sample(
  mspline,
  coefs_mean = NULL,
  prior_smooth = p_gamma(2, 1),
  prior_loghaz,
  prior_loghr = NULL,
  x = NULL,
  X = NULL,
  nonprop = FALSE,
  prior_sdnp = p_gamma(2, 1),
  nsim = 100
)

Arguments

mspline

A list of control parameters defining the spline model.

iknots: Internal knots. If this is not supplied, then the number of knots is taken from df, and their location is taken from equally-spaced quantiles of the observed event times (concatenated with the distinct follow-up times in the external data).

bknots: Boundary knots. If this is not supplied, the boundary knots are set to zero and the maximum event time (including the follow-up times in the external data).

df: Degrees of freedom, i.e. the number of parameters (or basis terms) that define the hazard as a spline function of time. Defaults to 10. Note this does not necessarily overfit because the function is smoothed through the prior.

degree: Polynomial degree used for the basis function. The default is 3, giving a cubic.

coefs_mean

Spline basis coefficients that define the prior mean for the hazard function. By default, these are set to values that define a constant hazard function. They are normalised to sum to 1 internally (if they do not already).

prior_smooth

Gamma prior for the smoothing standard deviation, in models where this is estimated. This should be a call to p_gamma(). The default is p_gamma(2,1).

prior_loghaz

Prior for the baseline log rate parameter (log(eta)), after centering around the log crude event rate in the data. This should be a call to a prior constructor function, such as p_normal(0,1) or p_t(0,2,2). Supported prior distribution families are normal (parameters mean and SD) and t distributions (parameters location, scale and degrees of freedom). The default is a normal distribution with mean 0 and standard deviation 20.

"Baseline" is defined by the continuous covariates taking a value of zero and factor covariates taking their reference level. To use a different baseline, the data should be transformed appropriately beforehand, so that a value of zero has a different meaning.

For continuous covariates, it helps for both computation and interpretation to define the value of zero to denote a typical value in the data, e.g. the mean.

prior_loghr

Priors for log hazard ratios. This should be a call to p_normal() or p_t(). A list of calls can also be provided, to give different priors to different coefficients, where the name of each list component matches the name of the coefficient, e.g. list("age45-59" = p_normal(0,1), "age60+" = p_t(0,2,3))

The default is p_normal(0,2.5) for all coefficients.

x

Only required for models with covariates. A list with components ncovs (number of covariates) and xnames (names of these covariates).

X

Either a list, vector or one-row matrix with one covariate value. Factors must be supplied as multiple numeric (0/1) contrasts. (TODO doc on how to build the names - vignette)

nonprop

Is the model a non-proportional hazards model (TRUE or FALSE).

prior_sdnp

Prior for the standard deviation parameters that smooth the non-proportionality effects over time in non-proportional hazards models. This should be a call to p_gamma() or a list of calls to p_gamma() with one component per covariate, as in prior_loghr.

nsim

Number of simulations to draw

Value

A list with components:

loghaz: Baseline log scale parameter (log(eta) in the notation of the manual). For models with covariates, this is at the covariate values supplied in X, or at zero if X is not supplied.

haz: Baseline scale parameter (eta).

coefs: Spline coefficients. For non-proportional hazards model with covariates, these are returned at the suppled value of X, or at values of zero if X is not supplied.

beta: Multinomial logit-transformed spline coefficients.

smooth: Smoothing standard deviation for spline coefficients.

Details

Cure and relative survival models are not currently handled.