Prior distributions and options

The functions described on this page are used to specify the prior distributions for the parameters in a survextrap model.

Usage

p_normal(location = 0, scale = 2.5)

p_t(location = 0, scale = 2.5, df = 1)

p_beta(shape1 = 1, shape2 = 1)

p_gamma(shape = 2, rate = 1)

Arguments

location

Prior location. For the normal distribution, this is the mean. Defaults to 0

scale

Prior scale. For the normal distribution, this is the standard deviation. Defaults to 2.5.

df

Prior degrees of freedom (only for Student t distribution).

shape1

First shape parameter (for Beta distribution, defaults to 1).

shape2

Second shape parameter (for Beta distribution, defaults to 1).

shape

Shape parameter (for Gamma distribution, defaults to 2).

rate

Rate parameter (for Gamma distribution, defaults to 1).

Value

A named list to be used internally by the survextrap model fitting functions.

See also