Posterior summary statistics for parameters of survextrap models

Posterior summary statistics for parameters of survextrap models. The summary statistics presented by default include the posterior median and 95% credible intervals, alongside the Rhat convergence diagnostic and the bulk effective sample size (as computed by the posterior package). For models fitted by optimisation rather than MCMC, the posterior mode is always returned.

Any other posterior summary can be computed if the appropriate function to compute it is supplied in summ_fns.

Usage

# S3 method for class 'survextrap'
summary(object, summ_fns = NULL, ...)

Arguments

object

A fitted model object as returned by survextrap

summ_fns

A list of functions to calculate different posterior summaries from the MCMC sample. This is passed to posterior::summarise_draws. If the list is named, then the names will be used for the columns of the output.

See the examples below for different ways this can be used.

Defaults to list(median = median, ~quantile(.x, probs=c(0.025, 0.975)), sd = sd, rhat = posterior::rhat, ess_bulk = posterior::ess_bulk)

Many useful such functions are provided with the posterior package.

...

Summary functions can also be supplied in separate arguments here. They will then be added to those supplied in summ_fns.

Value

A data frame (actually a tibble) of summary statistics for the model parameters.

The parameters, as indicated in the variable column, are:

alpha: Baseline log hazard scale. If there are covariates, this describes the log hazard scale with continuous covariates set to zero, and factor covariates set to their baseline levels. Note that this is not the log hazard, which also depends on the spline coefficients and basis. See hazard to extract the actual hazard.

coefs: Coefficients of the M-spline basis terms. If a non-proportional hazards model was fitted, these are with covariates set to zero or baseline levels.

loghr: Log hazard ratios for each covariate in the model. For cure models, this refers to covariates on survival for uncured people. For non-proportional hazards models, these are the multiplicative effects of covariates on the hazard scale parameter. See the methods vignette for a full description of this model.

hr: Hazard ratios (the exponentials of loghr).

pcure: Probability of cure (for cure models only). If there are covariates on cure, this parameter describes the probability of cure with continuous covariates set to zero, and factor covariates set to their baseline levels.

logor_cure: Log odds ratio of cure for each covariate on cure.

or_cure: Odds ratios of cure (the exponentials of logor_cure).

nperr: Standardised departures from proportional hazards in the non-proportional hazards model, defined as \(b^{(np)}_{ks} / \sigma^{(np)}_s\) (see the methods vignette for definitions of these).

hrsd: Smoothness standard deviations \(\tau_s\) for the non-proportionality effects.

Examples

mod <- survextrap(Surv(years, status) ~ rx, data=colons, fit_method="opt")
summary(mod)
#> # A tibble: 16 × 10
#>    variable basis_num term          mode   median     lower   upper     sd  rhat
#>    <chr>        <dbl> <chr>        <dbl>    <dbl>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 alpha           NA NA        -0.250   -0.249   -0.572     0.0779 0.164   1.00
#>  2 coefs            1 NA         0.00710  0.00718  0.000103  0.0382 0.0121  1.00
#>  3 coefs            2 NA         0.0175   0.0181   0.000468  0.0823 0.0380  1.00
#>  4 coefs            3 NA         0.0897   0.0772   0.0252    0.300  0.0727  1.00
#>  5 coefs            4 NA         0.119    0.104    0.0129    0.250  0.0562  1.00
#>  6 coefs            5 NA         0.114    0.0984   0.0231    0.237  0.0537  1.00
#>  7 coefs            6 NA         0.112    0.100    0.0224    0.255  0.0605  1.00
#>  8 coefs            7 NA         0.0692   0.0664   0.00441   0.215  0.0551  1.00
#>  9 coefs            8 NA         0.136    0.123    0.0272    0.287  0.0648  1.00
#> 10 coefs            9 NA         0.188    0.176    0.0470    0.375  0.0786  1.00
#> 11 coefs           10 NA         0.148    0.149    0.0400    0.328  0.0836  1.00
#> 12 loghr           NA rxLev     -0.384   -0.383   -0.901     0.117  0.256   1.00
#> 13 loghr           NA rxLev+5FU -0.664   -0.664   -1.20     -0.122  0.279   1.00
#> 14 hr              NA rxLev      0.681    0.682    0.406     1.12   0.185   1.00
#> 15 hr              NA rxLev+5FU  0.515    0.515    0.302     0.885  0.152   1.00
#> 16 hsd             NA NA         3.04     3.05     1.13      8.64   1.93    1.00
#> # ℹ 1 more variable: ess_bulk <dbl>
summary(mod, mean=mean)
#> # A tibble: 16 × 11
#>    variable basis_num term          mode   median     lower   upper     sd  rhat
#>    <chr>        <dbl> <chr>        <dbl>    <dbl>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 alpha           NA NA        -0.250   -0.249   -0.572     0.0779 0.164   1.00
#>  2 coefs            1 NA         0.00710  0.00718  0.000103  0.0382 0.0121  1.00
#>  3 coefs            2 NA         0.0175   0.0181   0.000468  0.0823 0.0380  1.00
#>  4 coefs            3 NA         0.0897   0.0772   0.0252    0.300  0.0727  1.00
#>  5 coefs            4 NA         0.119    0.104    0.0129    0.250  0.0562  1.00
#>  6 coefs            5 NA         0.114    0.0984   0.0231    0.237  0.0537  1.00
#>  7 coefs            6 NA         0.112    0.100    0.0224    0.255  0.0605  1.00
#>  8 coefs            7 NA         0.0692   0.0664   0.00441   0.215  0.0551  1.00
#>  9 coefs            8 NA         0.136    0.123    0.0272    0.287  0.0648  1.00
#> 10 coefs            9 NA         0.188    0.176    0.0470    0.375  0.0786  1.00
#> 11 coefs           10 NA         0.148    0.149    0.0400    0.328  0.0836  1.00
#> 12 loghr           NA rxLev     -0.384   -0.383   -0.901     0.117  0.256   1.00
#> 13 loghr           NA rxLev+5FU -0.664   -0.664   -1.20     -0.122  0.279   1.00
#> 14 hr              NA rxLev      0.681    0.682    0.406     1.12   0.185   1.00
#> 15 hr              NA rxLev+5FU  0.515    0.515    0.302     0.885  0.152   1.00
#> 16 hsd             NA NA         3.04     3.05     1.13      8.64   1.93    1.00
#> # ℹ 2 more variables: ess_bulk <dbl>, mean <dbl>
summary(mod, list(mean=mean))
#> # A tibble: 16 × 5
#>    variable basis_num term          mode    mean
#>    <chr>        <dbl> <chr>        <dbl>   <dbl>
#>  1 alpha           NA NA        -0.250   -0.250 
#>  2 coefs            1 NA         0.00710  0.0102
#>  3 coefs            2 NA         0.0175   0.0253
#>  4 coefs            3 NA         0.0897   0.0968
#>  5 coefs            4 NA         0.119    0.109 
#>  6 coefs            5 NA         0.114    0.105 
#>  7 coefs            6 NA         0.112    0.110 
#>  8 coefs            7 NA         0.0692   0.0749
#>  9 coefs            8 NA         0.136    0.132 
#> 10 coefs            9 NA         0.188    0.181 
#> 11 coefs           10 NA         0.148    0.157 
#> 12 loghr           NA rxLev     -0.384   -0.384 
#> 13 loghr           NA rxLev+5FU -0.664   -0.664 
#> 14 hr              NA rxLev      0.681    0.704 
#> 15 hr              NA rxLev+5FU  0.515    0.535 
#> 16 hsd             NA NA         3.04     3.48  
summary(mod, list(mean=mean, ess_tail=posterior::ess_tail))
#> # A tibble: 16 × 6
#>    variable basis_num term          mode    mean ess_tail
#>    <chr>        <dbl> <chr>        <dbl>   <dbl>    <dbl>
#>  1 alpha           NA NA        -0.250   -0.250     1919.
#>  2 coefs            1 NA         0.00710  0.0102    1951.
#>  3 coefs            2 NA         0.0175   0.0253    1817.
#>  4 coefs            3 NA         0.0897   0.0968    1881.
#>  5 coefs            4 NA         0.119    0.109     1776.
#>  6 coefs            5 NA         0.114    0.105     1987.
#>  7 coefs            6 NA         0.112    0.110     1812.
#>  8 coefs            7 NA         0.0692   0.0749    1921.
#>  9 coefs            8 NA         0.136    0.132     2007.
#> 10 coefs            9 NA         0.188    0.181     2002.
#> 11 coefs           10 NA         0.148    0.157     2047.
#> 12 loghr           NA rxLev     -0.384   -0.384     1666.
#> 13 loghr           NA rxLev+5FU -0.664   -0.664     1777.
#> 14 hr              NA rxLev      0.681    0.704     1666.
#> 15 hr              NA rxLev+5FU  0.515    0.535     1777.
#> 16 hsd             NA NA         3.04     3.48      1587.
summary(mod, mean=mean, ess_tail=posterior::ess_tail)
#> # A tibble: 16 × 12
#>    variable basis_num term          mode   median     lower   upper     sd  rhat
#>    <chr>        <dbl> <chr>        <dbl>    <dbl>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 alpha           NA NA        -0.250   -0.249   -0.572     0.0779 0.164   1.00
#>  2 coefs            1 NA         0.00710  0.00718  0.000103  0.0382 0.0121  1.00
#>  3 coefs            2 NA         0.0175   0.0181   0.000468  0.0823 0.0380  1.00
#>  4 coefs            3 NA         0.0897   0.0772   0.0252    0.300  0.0727  1.00
#>  5 coefs            4 NA         0.119    0.104    0.0129    0.250  0.0562  1.00
#>  6 coefs            5 NA         0.114    0.0984   0.0231    0.237  0.0537  1.00
#>  7 coefs            6 NA         0.112    0.100    0.0224    0.255  0.0605  1.00
#>  8 coefs            7 NA         0.0692   0.0664   0.00441   0.215  0.0551  1.00
#>  9 coefs            8 NA         0.136    0.123    0.0272    0.287  0.0648  1.00
#> 10 coefs            9 NA         0.188    0.176    0.0470    0.375  0.0786  1.00
#> 11 coefs           10 NA         0.148    0.149    0.0400    0.328  0.0836  1.00
#> 12 loghr           NA rxLev     -0.384   -0.383   -0.901     0.117  0.256   1.00
#> 13 loghr           NA rxLev+5FU -0.664   -0.664   -1.20     -0.122  0.279   1.00
#> 14 hr              NA rxLev      0.681    0.682    0.406     1.12   0.185   1.00
#> 15 hr              NA rxLev+5FU  0.515    0.515    0.302     0.885  0.152   1.00
#> 16 hsd             NA NA         3.04     3.05     1.13      8.64   1.93    1.00
#> # ℹ 3 more variables: ess_bulk <dbl>, mean <dbl>, ess_tail <dbl>