Posterior summary statistics for parameters of survextrap models

The summary statistics presented by default include the posterior median and 95% credible intervals, alongside the Rhat convergence diagnostic and the bulk effective sample size (as computed by the posterior package). For models fitted by optimisation rather than MCMC, the posterior mode is always returned.

Usage

# S3 method for survextrap
summary(object, summ_fns = NULL, ...)

Arguments

object

A fitted model object as returned by survextrap

summ_fns

A list of functions to calculate different posterior summaries from the MCMC sample. This is passed to posterior::summarise_draws. If the list is named, then the names will be used for the columns of the output.

See the examples below for different ways this can be used.

Defaults to list(median = median, ~quantile(.x, probs=c(0.025, 0.975)), sd = sd, rhat = posterior::rhat, ess_bulk = posterior::ess_bulk)

Many useful such functions are provided with the posterior package.

...

Summary functions can also be supplied in separate arguments here. They will then be added to those supplied in summ_fns.

Value

A data frame (actually a tibble) of summary statistics for the model parameters.

The parameters, as indicated in the variable column, are:

alpha: Baseline log hazard scale. If there are covariates, this describes the log hazard scale with continuous covariates set to zero, and factor covariates set to their baseline levels. Note that this is not the log hazard, which also depends on the spline coefficients and basis. See hazard to extract the actual hazard.

coefs: Coefficients of the M-spline basis terms. If a non-proportional hazards model was fitted, these are with covariates set to zero or baseline levels.

loghr: Log hazard ratios for each covariate in the model. For cure models, this refers to covariates on survival for uncured people. For non-proportional hazards models, these are the multiplicative effects of covariates on the hazard scale parameter. See the methods vignette

for a full description of this model.

hr: Hazard ratios (the exponentials of loghr).

pcure: Probability of cure (for cure models only). If there are covariates on cure, this parameter describes the probability of cure with continuous covariates set to zero, and factor covariates set to their baseline levels.

logor_cure: Log odds ratio of cure for each covariate on cure.

or_cure: Odds ratios of cure (the exponentials of logor_cure).

nperr: Standardised departures from proportional hazards in the non-proportional hazards model, defined as \(b^{(np)}_{ks} / \sigma^{(np)}_s\) (see the methods vignette

for definitions of these).

hrsd: Smoothness standard deviations \(\tau_s\) for the non-proportionality effects.

Details

Any other posterior summary can be computed if the appropriate function to compute it is supplied in summ_fns.

Examples

mod <- survextrap(Surv(years, status) ~ rx, data=colons, fit_method="opt")
summary(mod)
#> # A tibble: 16 × 10
#>    variable basis_num term          mode   median     lower   upper     sd  rhat
#>    <chr>        <dbl> <chr>        <dbl>    <dbl>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 alpha           NA NA        -0.255   -0.244    -5.66e-1  0.0581 0.158   1.00
#>  2 coefs            1 NA         0.00856  0.00729   3.07e-5  0.0350 0.0146  1.00
#>  3 coefs            2 NA         0.00414  0.00442   8.78e-7  0.408  0.126   1.00
#>  4 coefs            3 NA         0.129    0.0985    2.45e-3  0.235  0.0580  1.00
#>  5 coefs            4 NA         0.0865   0.0703    1.84e-3  0.342  0.0956  1.00
#>  6 coefs            5 NA         0.123    0.0947    2.98e-3  0.233  0.0570  1.00
#>  7 coefs            6 NA         0.140    0.115     1.90e-3  0.295  0.0761  1.00
#>  8 coefs            7 NA         0.0215   0.0221    6.81e-6  0.864  0.205   1.00
#>  9 coefs            8 NA         0.160    0.129     3.30e-3  0.307  0.0749  1.00
#> 10 coefs            9 NA         0.186    0.152     6.14e-3  0.361  0.0897  1.00
#> 11 coefs           10 NA         0.142    0.130     6.81e-3  0.387  0.0987  1.00
#> 12 loghr           NA rxLev     -0.383   -0.398    -8.93e-1  0.135  0.259   1.00
#> 13 loghr           NA rxLev+5FU -0.662   -0.663    -1.21e+0 -0.113  0.276   1.00
#> 14 hr              NA rxLev      0.682    0.672     4.09e-1  1.14   0.187   1.00
#> 15 hr              NA rxLev+5FU  0.516    0.515     2.99e-1  0.893  0.151   1.00
#> 16 hsd             NA NA         2.46     2.49      8.74e-1  6.78   1.62    1.00
#> # ℹ 1 more variable: ess_bulk <dbl>
summary(mod, mean=mean)
#> # A tibble: 16 × 11
#>    variable basis_num term          mode   median     lower   upper     sd  rhat
#>    <chr>        <dbl> <chr>        <dbl>    <dbl>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 alpha           NA NA        -0.255   -0.244    -5.66e-1  0.0581 0.158   1.00
#>  2 coefs            1 NA         0.00856  0.00729   3.07e-5  0.0350 0.0146  1.00
#>  3 coefs            2 NA         0.00414  0.00442   8.78e-7  0.408  0.126   1.00
#>  4 coefs            3 NA         0.129    0.0985    2.45e-3  0.235  0.0580  1.00
#>  5 coefs            4 NA         0.0865   0.0703    1.84e-3  0.342  0.0956  1.00
#>  6 coefs            5 NA         0.123    0.0947    2.98e-3  0.233  0.0570  1.00
#>  7 coefs            6 NA         0.140    0.115     1.90e-3  0.295  0.0761  1.00
#>  8 coefs            7 NA         0.0215   0.0221    6.81e-6  0.864  0.205   1.00
#>  9 coefs            8 NA         0.160    0.129     3.30e-3  0.307  0.0749  1.00
#> 10 coefs            9 NA         0.186    0.152     6.14e-3  0.361  0.0897  1.00
#> 11 coefs           10 NA         0.142    0.130     6.81e-3  0.387  0.0987  1.00
#> 12 loghr           NA rxLev     -0.383   -0.398    -8.93e-1  0.135  0.259   1.00
#> 13 loghr           NA rxLev+5FU -0.662   -0.663    -1.21e+0 -0.113  0.276   1.00
#> 14 hr              NA rxLev      0.682    0.672     4.09e-1  1.14   0.187   1.00
#> 15 hr              NA rxLev+5FU  0.516    0.515     2.99e-1  0.893  0.151   1.00
#> 16 hsd             NA NA         2.46     2.49      8.74e-1  6.78   1.62    1.00
#> # ℹ 2 more variables: ess_bulk <dbl>, mean <dbl>
summary(mod, list(mean=mean))
#> # A tibble: 16 × 5
#>    variable basis_num term          mode    mean
#>    <chr>        <dbl> <chr>        <dbl>   <dbl>
#>  1 alpha           NA NA        -0.255   -0.246 
#>  2 coefs            1 NA         0.00856  0.0102
#>  3 coefs            2 NA         0.00414  0.0388
#>  4 coefs            3 NA         0.129    0.102 
#>  5 coefs            4 NA         0.0865   0.0916
#>  6 coefs            5 NA         0.123    0.0982
#>  7 coefs            6 NA         0.140    0.122 
#>  8 coefs            7 NA         0.0215   0.106 
#>  9 coefs            8 NA         0.160    0.133 
#> 10 coefs            9 NA         0.186    0.156 
#> 11 coefs           10 NA         0.142    0.143 
#> 12 loghr           NA rxLev     -0.383   -0.389 
#> 13 loghr           NA rxLev+5FU -0.662   -0.659 
#> 14 hr              NA rxLev      0.682    0.701 
#> 15 hr              NA rxLev+5FU  0.516    0.538 
#> 16 hsd             NA NA         2.46     2.84  
summary(mod, list(mean=mean, ess_tail=posterior::ess_tail))
#> # A tibble: 16 × 6
#>    variable basis_num term          mode    mean ess_tail
#>    <chr>        <dbl> <chr>        <dbl>   <dbl>    <dbl>
#>  1 alpha           NA NA        -0.255   -0.246     1819.
#>  2 coefs            1 NA         0.00856  0.0102    1842.
#>  3 coefs            2 NA         0.00414  0.0388    1962.
#>  4 coefs            3 NA         0.129    0.102     1963.
#>  5 coefs            4 NA         0.0865   0.0916    2004.
#>  6 coefs            5 NA         0.123    0.0982    1695.
#>  7 coefs            6 NA         0.140    0.122     1961.
#>  8 coefs            7 NA         0.0215   0.106     1813.
#>  9 coefs            8 NA         0.160    0.133     1784.
#> 10 coefs            9 NA         0.186    0.156     1882.
#> 11 coefs           10 NA         0.142    0.143     1849.
#> 12 loghr           NA rxLev     -0.383   -0.389     1774.
#> 13 loghr           NA rxLev+5FU -0.662   -0.659     1846.
#> 14 hr              NA rxLev      0.682    0.701     1774.
#> 15 hr              NA rxLev+5FU  0.516    0.538     1846.
#> 16 hsd             NA NA         2.46     2.84      1884.
summary(mod, mean=mean, ess_tail=posterior::ess_tail)
#> # A tibble: 16 × 12
#>    variable basis_num term          mode   median     lower   upper     sd  rhat
#>    <chr>        <dbl> <chr>        <dbl>    <dbl>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 alpha           NA NA        -0.255   -0.244    -5.66e-1  0.0581 0.158   1.00
#>  2 coefs            1 NA         0.00856  0.00729   3.07e-5  0.0350 0.0146  1.00
#>  3 coefs            2 NA         0.00414  0.00442   8.78e-7  0.408  0.126   1.00
#>  4 coefs            3 NA         0.129    0.0985    2.45e-3  0.235  0.0580  1.00
#>  5 coefs            4 NA         0.0865   0.0703    1.84e-3  0.342  0.0956  1.00
#>  6 coefs            5 NA         0.123    0.0947    2.98e-3  0.233  0.0570  1.00
#>  7 coefs            6 NA         0.140    0.115     1.90e-3  0.295  0.0761  1.00
#>  8 coefs            7 NA         0.0215   0.0221    6.81e-6  0.864  0.205   1.00
#>  9 coefs            8 NA         0.160    0.129     3.30e-3  0.307  0.0749  1.00
#> 10 coefs            9 NA         0.186    0.152     6.14e-3  0.361  0.0897  1.00
#> 11 coefs           10 NA         0.142    0.130     6.81e-3  0.387  0.0987  1.00
#> 12 loghr           NA rxLev     -0.383   -0.398    -8.93e-1  0.135  0.259   1.00
#> 13 loghr           NA rxLev+5FU -0.662   -0.663    -1.21e+0 -0.113  0.276   1.00
#> 14 hr              NA rxLev      0.682    0.672     4.09e-1  1.14   0.187   1.00
#> 15 hr              NA rxLev+5FU  0.516    0.515     2.99e-1  0.893  0.151   1.00
#> 16 hsd             NA NA         2.46     2.49      8.74e-1  6.78   1.62    1.00
#> # ℹ 3 more variables: ess_bulk <dbl>, mean <dbl>, ess_tail <dbl>