Posterior summary statistics for parameters of survextrap models

Posterior summary statistics for parameters of survextrap models. The summary statistics presented by default include the posterior median and 95% credible intervals, alongside the Rhat convergence diagnostic and the bulk effective sample size (as computed by the posterior package). For models fitted by optimisation rather than MCMC, the posterior mode is always returned.

Any other posterior summary can be computed if the appropriate function to compute it is supplied in summ_fns.

Usage

# S3 method for survextrap
summary(object, summ_fns = NULL, ...)

Arguments

object

A fitted model object as returned by survextrap

summ_fns

A list of functions to calculate different posterior summaries from the MCMC sample. This is passed to posterior::summarise_draws. If the list is named, then the names will be used for the columns of the output.

See the examples below for different ways this can be used.

Defaults to list(median = median, ~quantile(.x, probs=c(0.025, 0.975)), sd = sd, rhat = posterior::rhat, ess_bulk = posterior::ess_bulk)

Many useful such functions are provided with the posterior package.

...

Summary functions can also be supplied in separate arguments here. They will then be added to those supplied in summ_fns.

Value

A data frame (actually a tibble) of summary statistics for the model parameters.

The parameters, as indicated in the variable column, are:

alpha: Baseline log hazard scale. If there are covariates, this describes the log hazard scale with continuous covariates set to zero, and factor covariates set to their baseline levels. Note that this is not the log hazard, which also depends on the spline coefficients and basis. See hazard to extract the actual hazard.

coefs: Coefficients of the M-spline basis terms. If a non-proportional hazards model was fitted, these are with covariates set to zero or baseline levels.

loghr: Log hazard ratios for each covariate in the model. For cure models, this refers to covariates on survival for uncured people. For non-proportional hazards models, these are the multiplicative effects of covariates on the hazard scale parameter. See the methods vignette

for a full description of this model.

hr: Hazard ratios (the exponentials of loghr).

pcure: Probability of cure (for cure models only). If there are covariates on cure, this parameter describes the probability of cure with continuous covariates set to zero, and factor covariates set to their baseline levels.

logor_cure: Log odds ratio of cure for each covariate on cure.

or_cure: Odds ratios of cure (the exponentials of logor_cure).

nperr: Standardised departures from proportional hazards in the non-proportional hazards model, defined as \(b^{(np)}_{ks} / \sigma^{(np)}_s\) (see the methods vignette

for definitions of these).

hrsd: Smoothness standard deviations \(\tau_s\) for the non-proportionality effects.

Examples

mod <- survextrap(Surv(years, status) ~ rx, data=colons, fit_method="opt")
summary(mod)
#> # A tibble: 16 × 10
#>    variable basis_num term          mode   median     lower   upper     sd  rhat
#>    <chr>        <dbl> <chr>        <dbl>    <dbl>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 alpha           NA NA        -0.254   -0.252    -5.70e-1  0.0749 0.161   1.00
#>  2 coefs            1 NA         0.00855  0.00776   4.43e-5  0.0356 0.0133  1.00
#>  3 coefs            2 NA         0.00414  0.00448   8.36e-7  0.515  0.135   1.00
#>  4 coefs            3 NA         0.129    0.0980    2.94e-3  0.238  0.0582  1.00
#>  5 coefs            4 NA         0.0865   0.0698    2.06e-3  0.323  0.0888  1.00
#>  6 coefs            5 NA         0.123    0.0952    2.43e-3  0.251  0.0627  1.00
#>  7 coefs            6 NA         0.140    0.113     3.29e-3  0.319  0.0782  1.00
#>  8 coefs            7 NA         0.0215   0.0202    5.68e-6  0.878  0.205   1.00
#>  9 coefs            8 NA         0.160    0.131     3.84e-3  0.308  0.0750  1.00
#> 10 coefs            9 NA         0.186    0.153     8.73e-3  0.338  0.0840  1.00
#> 11 coefs           10 NA         0.142    0.130     8.36e-3  0.355  0.0906  1.00
#> 12 loghr           NA rxLev     -0.383   -0.379    -8.85e-1  0.120  0.260   1.00
#> 13 loghr           NA rxLev+5FU -0.662   -0.650    -1.19e+0 -0.110  0.279   1.00
#> 14 hr              NA rxLev      0.682    0.684     4.13e-1  1.13   0.185   1.00
#> 15 hr              NA rxLev+5FU  0.516    0.522     3.05e-1  0.896  0.155   1.00
#> 16 hsd             NA NA         2.46     2.45      8.27e-1  7.07   1.63    1.00
#> # ℹ 1 more variable: ess_bulk <dbl>
summary(mod, mean=mean)
#> # A tibble: 16 × 11
#>    variable basis_num term          mode   median     lower   upper     sd  rhat
#>    <chr>        <dbl> <chr>        <dbl>    <dbl>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 alpha           NA NA        -0.254   -0.252    -5.70e-1  0.0749 0.161   1.00
#>  2 coefs            1 NA         0.00855  0.00776   4.43e-5  0.0356 0.0133  1.00
#>  3 coefs            2 NA         0.00414  0.00448   8.36e-7  0.515  0.135   1.00
#>  4 coefs            3 NA         0.129    0.0980    2.94e-3  0.238  0.0582  1.00
#>  5 coefs            4 NA         0.0865   0.0698    2.06e-3  0.323  0.0888  1.00
#>  6 coefs            5 NA         0.123    0.0952    2.43e-3  0.251  0.0627  1.00
#>  7 coefs            6 NA         0.140    0.113     3.29e-3  0.319  0.0782  1.00
#>  8 coefs            7 NA         0.0215   0.0202    5.68e-6  0.878  0.205   1.00
#>  9 coefs            8 NA         0.160    0.131     3.84e-3  0.308  0.0750  1.00
#> 10 coefs            9 NA         0.186    0.153     8.73e-3  0.338  0.0840  1.00
#> 11 coefs           10 NA         0.142    0.130     8.36e-3  0.355  0.0906  1.00
#> 12 loghr           NA rxLev     -0.383   -0.379    -8.85e-1  0.120  0.260   1.00
#> 13 loghr           NA rxLev+5FU -0.662   -0.650    -1.19e+0 -0.110  0.279   1.00
#> 14 hr              NA rxLev      0.682    0.684     4.13e-1  1.13   0.185   1.00
#> 15 hr              NA rxLev+5FU  0.516    0.522     3.05e-1  0.896  0.155   1.00
#> 16 hsd             NA NA         2.46     2.45      8.27e-1  7.07   1.63    1.00
#> # ℹ 2 more variables: ess_bulk <dbl>, mean <dbl>
summary(mod, list(mean=mean))
#> # A tibble: 16 × 5
#>    variable basis_num term          mode    mean
#>    <chr>        <dbl> <chr>        <dbl>   <dbl>
#>  1 alpha           NA NA        -0.254   -0.256 
#>  2 coefs            1 NA         0.00855  0.0107
#>  3 coefs            2 NA         0.00414  0.0450
#>  4 coefs            3 NA         0.129    0.102 
#>  5 coefs            4 NA         0.0865   0.0890
#>  6 coefs            5 NA         0.123    0.100 
#>  7 coefs            6 NA         0.140    0.122 
#>  8 coefs            7 NA         0.0215   0.105 
#>  9 coefs            8 NA         0.160    0.135 
#> 10 coefs            9 NA         0.186    0.154 
#> 11 coefs           10 NA         0.142    0.138 
#> 12 loghr           NA rxLev     -0.383   -0.382 
#> 13 loghr           NA rxLev+5FU -0.662   -0.650 
#> 14 hr              NA rxLev      0.682    0.706 
#> 15 hr              NA rxLev+5FU  0.516    0.543 
#> 16 hsd             NA NA         2.46     2.83  
summary(mod, list(mean=mean, ess_tail=posterior::ess_tail))
#> # A tibble: 16 × 6
#>    variable basis_num term          mode    mean ess_tail
#>    <chr>        <dbl> <chr>        <dbl>   <dbl>    <dbl>
#>  1 alpha           NA NA        -0.254   -0.256     1728.
#>  2 coefs            1 NA         0.00855  0.0107    2003.
#>  3 coefs            2 NA         0.00414  0.0450    1719.
#>  4 coefs            3 NA         0.129    0.102     2040.
#>  5 coefs            4 NA         0.0865   0.0890    2004.
#>  6 coefs            5 NA         0.123    0.100     1852.
#>  7 coefs            6 NA         0.140    0.122     1926.
#>  8 coefs            7 NA         0.0215   0.105     1920.
#>  9 coefs            8 NA         0.160    0.135     1984.
#> 10 coefs            9 NA         0.186    0.154     1896.
#> 11 coefs           10 NA         0.142    0.138     1885.
#> 12 loghr           NA rxLev     -0.383   -0.382     1927.
#> 13 loghr           NA rxLev+5FU -0.662   -0.650     1749.
#> 14 hr              NA rxLev      0.682    0.706     1927.
#> 15 hr              NA rxLev+5FU  0.516    0.543     1749.
#> 16 hsd             NA NA         2.46     2.83      1922.
summary(mod, mean=mean, ess_tail=posterior::ess_tail)
#> # A tibble: 16 × 12
#>    variable basis_num term          mode   median     lower   upper     sd  rhat
#>    <chr>        <dbl> <chr>        <dbl>    <dbl>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 alpha           NA NA        -0.254   -0.252    -5.70e-1  0.0749 0.161   1.00
#>  2 coefs            1 NA         0.00855  0.00776   4.43e-5  0.0356 0.0133  1.00
#>  3 coefs            2 NA         0.00414  0.00448   8.36e-7  0.515  0.135   1.00
#>  4 coefs            3 NA         0.129    0.0980    2.94e-3  0.238  0.0582  1.00
#>  5 coefs            4 NA         0.0865   0.0698    2.06e-3  0.323  0.0888  1.00
#>  6 coefs            5 NA         0.123    0.0952    2.43e-3  0.251  0.0627  1.00
#>  7 coefs            6 NA         0.140    0.113     3.29e-3  0.319  0.0782  1.00
#>  8 coefs            7 NA         0.0215   0.0202    5.68e-6  0.878  0.205   1.00
#>  9 coefs            8 NA         0.160    0.131     3.84e-3  0.308  0.0750  1.00
#> 10 coefs            9 NA         0.186    0.153     8.73e-3  0.338  0.0840  1.00
#> 11 coefs           10 NA         0.142    0.130     8.36e-3  0.355  0.0906  1.00
#> 12 loghr           NA rxLev     -0.383   -0.379    -8.85e-1  0.120  0.260   1.00
#> 13 loghr           NA rxLev+5FU -0.662   -0.650    -1.19e+0 -0.110  0.279   1.00
#> 14 hr              NA rxLev      0.682    0.684     4.13e-1  1.13   0.185   1.00
#> 15 hr              NA rxLev+5FU  0.516    0.522     3.05e-1  0.896  0.155   1.00
#> 16 hsd             NA NA         2.46     2.45      8.27e-1  7.07   1.63    1.00
#> # ℹ 3 more variables: ess_bulk <dbl>, mean <dbl>, ess_tail <dbl>