Posterior summary statistics for parameters of survextrap models

By default, a median and 95% credible intervals are presented, alongside the Rhat convergence diagnostic and the bulk effective sample size (see the posterior package). For models fitted by optimisation rather than MCMC, the posterior mode is also returned.

Usage

# S3 method for survextrap
summary(object, ...)

Arguments

object

A fitted model object as returned by survextrap

...

Other functions to compute a summary statistic from a vector of posterior samples can be supplied here. Many useful such functions are provided with the posterior package. If the arguments are named, then these are used to name the returned data frame. See the examples below.

Value

A data frame (actually a tibble) of summary statistics for the model parameters is returned.

The parameters, as indicated in the variable column, are:

alpha: Average baseline log hazard. If there are covariates, this describes the average hazard with continuous covariates set to zero, and factor covariates set to their baseline levels.

coefs: Coefficients of the M-spline basis terms. If a non-proportional hazards model was fitted, these are with covariates set to zero or baseline levels.

loghr: Log hazard ratios for each covariate in the model. For cure models, this refers to covariates on survival for uncured people. For non-proportional hazards models, these are the effects of covariates on the scale parameter eta and represent "average" hazard ratios, from which departures are modelled (see the "methods" vignette for a full description of this model).

hr: Hazard ratios (the exponentials of loghr).

pcure: Probability of cure (for cure models only). If there are covariates on cure, this parameter describes the probability of cure with continuous covariates set to zero, and factor covariates set to their baseline levels.

logor_cure: Log odds ratio of cure for each covariate on cure.

or_cure: Odds ratios of cure (the exponentials of logor_cure).

nperr: Standardised departures from proportional hazards in the non-proportional hazards model, defined as \(b^{(np)}_{ks} / \sigma^{(np)}_s\) (see the "methods" vignette for definitions of these).

sd_np: Smoothness standard deviations \(\sigma^{(np)}_s\) for the non-proportionality effects.

Details

Any other posterior summary can be computed if the appropriate function to compute it is supplied here.

Examples

mod <- survextrap(Surv(years, status) ~ rx, data=colons, fit_method="opt")
summary(mod)
#> # A tibble: 16 × 10
#>    variable  basis_num term         mode   median    lower   upper      sd  rhat
#>    <chr>         <dbl> <chr>       <dbl>    <dbl>    <dbl>   <dbl>   <dbl> <dbl>
#>  1 alpha            NA NA       -0.267   -0.269   -5.80e-1  0.0449 0.159    1.00
#>  2 coefs             1 NA        0.00974  0.00721  1.04e-5  0.0313 0.00936  1.00
#>  3 coefs             2 NA        0.0118   0.0101   1.21e-6  0.706  0.167    1.00
#>  4 coefs             3 NA        0.156    0.108    3.53e-4  0.276  0.0712   1.00
#>  5 coefs             4 NA        0.137    0.0930   7.60e-4  0.247  0.0667   1.00
#>  6 coefs             5 NA        0.178    0.128    3.21e-4  0.317  0.0859   1.00
#>  7 coefs             6 NA        0.0348   0.0270   2.53e-5  0.867  0.209    1.00
#>  8 coefs             7 NA        0.287    0.213    1.07e-3  0.419  0.114    1.00
#>  9 coefs             8 NA        0.141    0.106    8.85e-4  0.489  0.123    1.00
#> 10 coefs             9 NA        0.0235   0.0205   1.68e-6  0.834  0.195    1.00
#> 11 coefs            10 NA        0.0223   0.0181   1.22e-4  0.174  0.0795   1.00
#> 12 loghr            NA rxLev    -0.382   -0.381   -8.66e-1  0.117  0.257    1.00
#> 13 loghr            NA rxLev+5… -0.661   -0.666   -1.20e+0 -0.130  0.268    1.00
#> 14 hr               NA rxLev     0.683    0.683    4.20e-1  1.12   0.185    1.00
#> 15 hr               NA rxLev+5…  0.516    0.514    3.01e-1  0.878  0.146    1.00
#> 16 smooth_sd        NA NA        2.32     2.28     8.06e-1  6.87   1.58     1.00
#> # … with 1 more variable: ess_bulk <dbl>
summary(mod, mean)
#> # A tibble: 16 × 11
#>    variable  basis_num term         mode   median    lower   upper      sd  rhat
#>    <chr>         <dbl> <chr>       <dbl>    <dbl>    <dbl>   <dbl>   <dbl> <dbl>
#>  1 alpha            NA NA       -0.267   -0.269   -5.80e-1  0.0449 0.159    1.00
#>  2 coefs             1 NA        0.00974  0.00721  1.04e-5  0.0313 0.00936  1.00
#>  3 coefs             2 NA        0.0118   0.0101   1.21e-6  0.706  0.167    1.00
#>  4 coefs             3 NA        0.156    0.108    3.53e-4  0.276  0.0712   1.00
#>  5 coefs             4 NA        0.137    0.0930   7.60e-4  0.247  0.0667   1.00
#>  6 coefs             5 NA        0.178    0.128    3.21e-4  0.317  0.0859   1.00
#>  7 coefs             6 NA        0.0348   0.0270   2.53e-5  0.867  0.209    1.00
#>  8 coefs             7 NA        0.287    0.213    1.07e-3  0.419  0.114    1.00
#>  9 coefs             8 NA        0.141    0.106    8.85e-4  0.489  0.123    1.00
#> 10 coefs             9 NA        0.0235   0.0205   1.68e-6  0.834  0.195    1.00
#> 11 coefs            10 NA        0.0223   0.0181   1.22e-4  0.174  0.0795   1.00
#> 12 loghr            NA rxLev    -0.382   -0.381   -8.66e-1  0.117  0.257    1.00
#> 13 loghr            NA rxLev+5… -0.661   -0.666   -1.20e+0 -0.130  0.268    1.00
#> 14 hr               NA rxLev     0.683    0.683    4.20e-1  1.12   0.185    1.00
#> 15 hr               NA rxLev+5…  0.516    0.514    3.01e-1  0.878  0.146    1.00
#> 16 smooth_sd        NA NA        2.32     2.28     8.06e-1  6.87   1.58     1.00
#> # … with 2 more variables: ess_bulk <dbl>, mean <dbl>
summary(mod, mean, ess_tail=posterior::ess_tail)
#> # A tibble: 16 × 12
#>    variable  basis_num term         mode   median    lower   upper      sd  rhat
#>    <chr>         <dbl> <chr>       <dbl>    <dbl>    <dbl>   <dbl>   <dbl> <dbl>
#>  1 alpha            NA NA       -0.267   -0.269   -5.80e-1  0.0449 0.159    1.00
#>  2 coefs             1 NA        0.00974  0.00721  1.04e-5  0.0313 0.00936  1.00
#>  3 coefs             2 NA        0.0118   0.0101   1.21e-6  0.706  0.167    1.00
#>  4 coefs             3 NA        0.156    0.108    3.53e-4  0.276  0.0712   1.00
#>  5 coefs             4 NA        0.137    0.0930   7.60e-4  0.247  0.0667   1.00
#>  6 coefs             5 NA        0.178    0.128    3.21e-4  0.317  0.0859   1.00
#>  7 coefs             6 NA        0.0348   0.0270   2.53e-5  0.867  0.209    1.00
#>  8 coefs             7 NA        0.287    0.213    1.07e-3  0.419  0.114    1.00
#>  9 coefs             8 NA        0.141    0.106    8.85e-4  0.489  0.123    1.00
#> 10 coefs             9 NA        0.0235   0.0205   1.68e-6  0.834  0.195    1.00
#> 11 coefs            10 NA        0.0223   0.0181   1.22e-4  0.174  0.0795   1.00
#> 12 loghr            NA rxLev    -0.382   -0.381   -8.66e-1  0.117  0.257    1.00
#> 13 loghr            NA rxLev+5… -0.661   -0.666   -1.20e+0 -0.130  0.268    1.00
#> 14 hr               NA rxLev     0.683    0.683    4.20e-1  1.12   0.185    1.00
#> 15 hr               NA rxLev+5…  0.516    0.514    3.01e-1  0.878  0.146    1.00
#> 16 smooth_sd        NA NA        2.32     2.28     8.06e-1  6.87   1.58     1.00
#> # … with 3 more variables: ess_bulk <dbl>, mean <dbl>, ess_tail <dbl>