Flexible Bayesian parametric survival models

Flexible Bayesian parametric survival models. Individual data are represented using M-splines and a proportional hazards or flexible non-proportional hazards model. Optionally a mixture cure version of this model can be fitted. Extrapolations can be enhanced by including external aggregate data, or by including a fixed background hazard in an additive hazards (relative survival) model.

Usage

survextrap(
  formula,
  data,
  external = NULL,
  smooth_sd = "bayes",
  coefs_mean = NULL,
  cure = FALSE,
  nonprop = FALSE,
  prior_loghaz = NULL,
  prior_loghr = NULL,
  prior_smooth = p_gamma(2, 1),
  prior_cure = p_beta(1, 1),
  prior_logor_cure = NULL,
  prior_sdnp = p_gamma(2, 1),
  backhaz = NULL,
  basehaz_ops = NULL,
  fit_method = "mcmc",
  loo = (fit_method == "mcmc"),
  ...
)

Arguments

formula

A survival formula in standard R formula syntax, with a call to Surv() on the left hand side.

Covariates included on the right hand side of the formula with be modelled with proportional hazards.

data

Data frame containing variables in formula.

external

External data as a data frame of aggregate survival counts with columns:

start: Start time

start: Follow-up time

n: Number of people alive at start

r: Number of those people who are still alive at stop

It is intended to add facilities to the package to produce this format of data from other common forms of external data, e.g. registry or population data, or elicited judgements about conditional survival

smooth_sd

Smoothing parameter estimation.

"bayes": the smoothing parameter is estimated by full Bayes.

"eb": empirical Bayes is used.

Alternatively, if a number is supplied here, then the smoothing parameter is fixed to this number.

coefs_mean

Spline basis coefficients that define the prior mean for the hazard function. By default, these are set to values that define a constant hazard function. They are normalised to sum to 1 internally (if they do not already).

cure

If TRUE, a mixture cure model is used, where the "uncured" survival is defined by the M-spline model, and the cure probability is estimated.

nonprop

If TRUE then a non-proportional hazards model is fitted. This is achieved by modelling the spline basis weights in terms of the covariates. See the "methods" vignette for more details. In models with multiple covariates, currently there is no way to assume that some covariates have proportional hazards but others don't - it is all or none.

prior_loghaz

Prior for the baseline log hazard, after centering around the log crude event rate in the data. This should be a call to a prior constructor function, such as p_normal(0,1) or p_t(0,2,2). Supported prior distribution families are normal (parameters mean and SD) and t distributions (parameters location, scale and degrees of freedom). The default is a normal distribution with a mean defined by the log crude event rate in the data, and a standard deviation in the data.

"Baseline" is defined by the continuous covariates taking a value of zero and factor covariates taking their reference level. To use a different baseline, the data should be transformed appropriately beforehand, so that a value of zero has a different meaning.

For continuous covariates, it helps for both computation and interpretation to define the value of zero to denote a typical value in the data, e.g. the mean.

prior_loghr

Priors for log hazard ratios. This should be a call to p_normal() or p_t(). A list of calls can also be provided, to give different priors to different coefficients, where the name of each list component matches the name of the coefficient, e.g. list("age45-59" = p_normal(0,1), "age60+" = p_t(0,2,3))

The default is p_normal(0,2.5) for all coefficients.

prior_smooth

Gamma prior for the smoothing standard deviation, in models where this is estimated. This should be a call to p_gamma(). The default is p_gamma(2,1).

prior_cure

Prior for the baseline cure probability. This should be a call to p_beta(). The default is a uniform prior, p_beta(1,1). Baseline is defined by the mean of continuous covariates and the reference level of factor covariates.

prior_logor_cure

Priors for log odds ratios on cure probabilities. This should be a call to p_normal() or p_t(). The default is p_normal(0,2.5).

prior_sdnp

Prior for the standard deviation parameters that smooth the non-proportionality effects over time in non-proportional hazards models. This should be a call to p_gamma() or a list of calls to p_gamma() with one component per covariate, as in prior_loghr.

backhaz

Background hazard, that is, for causes of death other than the cause of interest. This defines a "relative survival" model where the overall hazard is the sum of a cause-specific hazard and a background hazard. The background hazard is assumed to be known, and the cause-specific hazard is modelled with the flexible parametric model.

The background hazard can be supplied in two forms. The meaning of predictions from the model depends on which of these is used.

(a) A data frame with columns "hazard" and "time", specifying the "background" hazard at all times as a piecewise-constant (step) function. Each row gives the "background" hazard between the specified time and the next time. The first element of "time" should be 0, and the final row specifies the hazard at all times greater than the last element of "time".

Predictions from the modelled fitted by survextrap will then include this background hazard, because it is known at all times.

(b) The (quoted) name of a variable in the data giving the background hazard. For censored cases, the exact value does not matter. The predictions from survextrap will then describe the excess hazard or survival on top of this background. The overall hazard cannot be predicted in general, because the background hazard is only specified over a limited range of time.

If there is external data and backhaz is supplied in form (b), then the user should also supply the background survival at the start and stop points in columns of the external data named "backsurv_start" and "backsurv_stop". This should describe the same reference population as backhaz, though the package does not check for consistency between these.

Leave-one-out cross-validation currently does not take into account any background hazard (effectively assuming it to be zero).

basehaz_ops

A list of control parameters defining the spline model.

iknots: Internal knots. If this is not supplied, then the number of knots is taken from df, and their location is taken from equally-spaced quantiles of the observed event times (concatenated with the distinct follow-up times in the external data).

bknots: Boundary knots. If this is not supplied, the boundary knots are set to zero and the maximum event time (including the follow-up times in the external data).

df: Degrees of freedom, i.e. the number of parameters (or basis terms) that define the hazard as a spline function of time. Defaults to 10. Note this does not necessarily overfit because the function is smoothed through the prior.

degree: Polynomial degree used for the basis function. The default is 3, giving a cubic.

fit_method

Method from rstan used to fit the model. Defaults to MCMC.

If fit_method="mcmc" then a sample from the posterior is drawn using Markov Chain Monte Carlo sampling, via rstan's rstan::sampling() function. This is the default. It is the most accurate but the slowest method.

If fit_method="opt", then instead of an MCMC sample from the posterior, survextrap returns the posterior mode calculated using optimisation, via rstan's rstan::optimizing() function. A sample from a normal approximation to the (real-line-transformed) posterior distribution is drawn in order to obtain credible intervals.

If fit_method="vb", then variational Bayes methods are used, via rstan's rstan::vb() function. This is labelled as "experimental" by rstan. It might give a better approximation to the posterior than fit_method="opt", but has not been investigated in depth for these models.

loo

Compute leave-one-out cross-validation statistics. This is done by default. Set to FALSE to not compute them. If these statistics are computed, then they are returned in the loo component of the object returned by survextrap. See the "examples" vignette for some explanation of these.

...

Additional arguments to supply to control the Stan fit, passed to the appropriate rstan function, depending on which is chosen through the fit_method argument.

Value

A list of objects defining the fitted model. A particularly important component is

stanfit: The object returned by rstan containing samples from the fitted model. See the rstan documentation.